+++ lib/doppler/action.js
@@ -9,7 +9,7 @@ const getX = require('./get-x.js');
 const gotoTime = require('./goto-time.js');
 const statusBox = require('../statusBox.js');
 const msgbox = require('../messageBox.js');
-const statusBarMgr = require('../statusBarMgr.js');
+const statusBarUpdater = require('../statusBarUpdater.js');
 
 const createBookmarkAtCursor = (event, content, cursor, pstate, ws, render) => {
   // Create a new bookmark from the PV.
@@ -167,18 +167,20 @@ const cursorsyncmode = (event, content, cursor, pstate, ws, render) => {
   // console.log("cursor sync mode");
 
   ws.sendcmd(['cursorsyncmode']);
-  statusBarMgr.toggleCurs();
+  statusBarUpdater.toggleCurs();
   return true;
 };
 
 const cursorSyncmode = (ws) => {
+  // console.log("cursor sync mode");
+
   ws.sendcmd(['cursorsyncmode']);
-  statusBarMgr.toggleNavs();
+  statusBarUpdater.toggleCurs();
   return true;
 };
 
 const togglenavsync = (event, content, cursor, pstate, ws, render) => {
-  statusBarMgr.toggleNavs();
+  statusBarUpdater.toggleNavs();
   ws.sendcmd(['togglenavsync']);
   return true;
 };
