+++ lib/statusBarMgr.js
@@ -2,6 +2,7 @@
 const statusBox = require('./statusBox.js');
 const action = require('./doppler/action.js');
 const msgbox = require('./messageBox.js');
+const statusBarUpdater = require('./statusBarUpdater.js');
 
 const updateStatusBoxVisibility = (pstate) => {
   let statusBar = document.getElementById('statusBarContainer');
@@ -58,45 +59,16 @@ let createStatusBars = (pstate, cm, ws) => {
 
 let toggleNavStatus = (ws) => {
   action.toggleNavsync(ws);
-  toggleNavs();
-};
-
-let toggleNavs = () => {
-  var x = document.getElementById('navsync');
-  const backgroundColor = x.style.backgroundColor;
-  if (x.innerHTML === 'NavSync: off') {
-    x.style.backgroundColor = 'rgb(262, 63, 78)';
-    x.innerHTML = 'NavSync: on';
-  } else {
-    x.innerHTML = 'NavSync: off';
-    x.style.backgroundColor = 'rgb(66, 66, 66)';
-  }
+  statusBarUpdater.toggleNavs();
 };
 
 let toggleCurStatus = (ws) => {
   action.cursorSyncmode(ws);
-  toggleCurs();
-};
-
-let toggleCurs = () => {
-  var x = document.getElementById('cursync');
-  const backgroundColor = x.style.backgroundColor;
-  if (x.innerHTML === 'CurSync: abs') {
-    x.style.backgroundColor = 'green';
-    x.innerHTML = 'CurSync: rel';
-  } else if (x.innerHTML === 'CurSync: rel') {
-    x.innerHTML = 'CurSync: off';
-    x.style.backgroundColor = 'rgb(66, 66, 66)';
-  } else {
-    x.innerHTML = 'CurSync: abs';
-    x.style.backgroundColor = 'rgb(262, 63, 78)';
-  }
+  statusBarUpdater.toggleCurs();
 };
 
 module.exports = {
   updateStatusBoxVisibility,
   createStatusBars,
   toggleNavStatus,
-  toggleCurs,
-  toggleNavs,
 };
