+++ lib/statusBarUpdater.js
@@ -1,0 +1,36 @@
+'use strict';
+// const statusBox = require('./statusBox.js');
+// const action = require('./doppler/action.js');
+// const msgbox = require('./messageBox.js');
+
+let toggleNavs = () => {
+  var x = document.getElementById('navsync');
+  const backgroundColor = x.style.backgroundColor;
+  if (x.innerHTML === 'NavSync: off') {
+    x.style.backgroundColor = 'rgb(262, 63, 78)';
+    x.innerHTML = 'NavSync: on';
+  } else {
+    x.innerHTML = 'NavSync: off';
+    x.style.backgroundColor = 'rgb(66, 66, 66)';
+  }
+};
+
+let toggleCurs = () => {
+  var x = document.getElementById('cursync');
+  const backgroundColor = x.style.backgroundColor;
+  if (x.innerHTML === 'CurSync: abs') {
+    x.style.backgroundColor = 'green';
+    x.innerHTML = 'CurSync: rel';
+  } else if (x.innerHTML === 'CurSync: rel') {
+    x.innerHTML = 'CurSync: off';
+    x.style.backgroundColor = 'rgb(66, 66, 66)';
+  } else {
+    x.innerHTML = 'CurSync: abs';
+    x.style.backgroundColor = 'rgb(262, 63, 78)';
+  }
+};
+
+module.exports = {
+  toggleCurs,
+  toggleNavs,
+};
